RNG PMacc: add 64Bit support
- fix miner documentation bugs
- add support for 64Bit floating point
  - normal distributed numbers
  - uniform distributed numbers
- add 64bit integer support to generate uniform distributed random numbers

+++ include/pmacc/random/distributions/normal/Normal_double.hpp
@@ -45,12 +45,12 @@ namespace detail
         typename T_Acc
     >
     struct Normal<
-        float,
+        double,
         methods::XorMin< T_Acc >,
         void
     > :
         public MullerBox<
-            float,
+            double,
             methods::XorMin< T_Acc >
         >
     {
@@ -62,12 +62,12 @@ namespace detail
         typename T_Acc
     >
     struct Normal<
-        float,
+        double,
         methods::MRG32k3aMin< T_Acc >,
         void
     > :
         public MullerBox<
-            float,
+            double,
             methods::MRG32k3aMin< T_Acc >
         >
     {
